atualiazado

+++ src/main/java/com/receitas/culin/rias/model/Usuario.java
@@ -17,7 +17,7 @@ public class Usuario implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String nome;
     private String email;
     private String senha;
@@ -30,18 +30,18 @@ public class Usuario implements UserDetails {
     }
 
     //getters e setters
-    public Usuario(int id, String nome, String email, String senha) {
+    public Usuario(long id, String nome, String email, String senha) {
         this.id = id;
         this.nome = nome;
         this.email = email;
         this.senha = senha;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
